chore: add licence disclaimers
